# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WRS Digital Agricultural Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserTier = typeof UserTier[keyof typeof UserTier];


export const UserTier = {
  PRODUCER: 'PRODUCER',
  OFF_TAKER: 'OFF_TAKER',
  ENABLER: 'ENABLER',
  FINANCIER: 'FINANCIER',
} as const;

export type UserKybStatus = typeof UserKybStatus[keyof typeof UserKybStatus];


export const UserKybStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED',
} as const;

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  tier: UserTier;
  reputationScore: number;
  kybStatus: UserKybStatus;
  /** @nullable */
  company?: string | null;
  createdAt: string;
}

export type UserUpdateTier = typeof UserUpdateTier[keyof typeof UserUpdateTier];


export const UserUpdateTier = {
  PRODUCER: 'PRODUCER',
  OFF_TAKER: 'OFF_TAKER',
  ENABLER: 'ENABLER',
  FINANCIER: 'FINANCIER',
} as const;

export interface UserUpdate {
  name?: string;
  company?: string;
  tier?: UserUpdateTier;
}

export type EwrCommodityType = typeof EwrCommodityType[keyof typeof EwrCommodityType];


export const EwrCommodityType = {
  MAIZE: 'MAIZE',
  RICE: 'RICE',
  COFFEE: 'COFFEE',
  TEA: 'TEA',
  AVOCADO: 'AVOCADO',
} as const;

export type EwrState = typeof EwrState[keyof typeof EwrState];


export const EwrState = {
  INGESTED: 'INGESTED',
  MARKET_LISTED: 'MARKET_LISTED',
  LOCK_TRADING: 'LOCK_TRADING',
  SETTLED: 'SETTLED',
  ENCUMBERED: 'ENCUMBERED',
} as const;

export interface Ewr {
  id: number;
  ewrsReceiptId: string;
  wrscSignature: string;
  warehouseCode: string;
  commodityType: EwrCommodityType;
  grade: string;
  weightMt: number;
  /** @nullable */
  moisturePct: number | null;
  harvestSeason: string;
  isLienActive: boolean;
  /** @nullable */
  lienHolderId?: number | null;
  state: EwrState;
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  expiryAt?: string | null;
  issuedAt: string;
  /** @nullable */
  estimatedValueUsd?: number | null;
}

export type PortfolioByStateItem = {
  state: string;
  count: number;
};

export type PortfolioByCommodityItem = {
  commodityType: string;
  count: number;
  totalWeightMt: number;
};

export interface Portfolio {
  ewrs: Ewr[];
  totalValueUsd: number;
  byState: PortfolioByStateItem[];
  byCommodity: PortfolioByCommodityItem[];
}

export type SpotListingStatus = typeof SpotListingStatus[keyof typeof SpotListingStatus];


export const SpotListingStatus = {
  ACTIVE: 'ACTIVE',
  LOCKED: 'LOCKED',
  SETTLED: 'SETTLED',
  CANCELLED: 'CANCELLED',
} as const;

export interface SpotListing {
  id: number;
  ewrId: number;
  sellerId: number;
  /** @nullable */
  sellerName?: string | null;
  pricePerMt: number;
  currency: string;
  status: SpotListingStatus;
  /** @nullable */
  commodityType?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  weightMt?: number | null;
  /** @nullable */
  warehouseCode?: string | null;
  /** @nullable */
  moisturePct?: number | null;
  /** @nullable */
  harvestSeason?: string | null;
  /** @nullable */
  totalValueUsd?: number | null;
  /** @nullable */
  platformFeeUsd?: number | null;
  /** @nullable */
  escrowFeeUsd?: number | null;
  createdAt: string;
}

export interface SpotListingDetail {
  listing: SpotListing;
  ewr: Ewr;
}

export interface SpotListingInput {
  ewrId: number;
  pricePerMt: number;
  currency: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  PENDING_SETTLEMENT: 'PENDING_SETTLEMENT',
  SETTLED: 'SETTLED',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED',
} as const;

export interface Order {
  id: number;
  listingId: number;
  buyerId: number;
  /** @nullable */
  buyerName?: string | null;
  lockedAt: string;
  expiresAt: string;
  /** @nullable */
  settledAt?: string | null;
  status: OrderStatus;
  totalUsd: number;
  platformFeeUsd: number;
  escrowFeeUsd: number;
  /** @nullable */
  commodityType?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  weightMt?: number | null;
  /** @nullable */
  warehouseCode?: string | null;
  /** @nullable */
  pricePerMt?: number | null;
  /** @nullable */
  sellerName?: string | null;
  createdAt: string;
}

export interface OrderInput {
  listingId: number;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  SETTLED: 'SETTLED',
  CANCELLED: 'CANCELLED',
} as const;

export interface OrderUpdate {
  status: OrderUpdateStatus;
}

export interface MarketSummary {
  totalActiveListings: number;
  totalEwrsIngested: number;
  totalEwrs: number;
  totalSettledToday: number;
  totalVolumeUsd: number;
  /** @nullable */
  avgPricePerMt?: number | null;
}

export interface CommodityStat {
  commodityType: string;
  activeListings: number;
  totalEwrs: number;
  totalWeightMt: number;
  /** @nullable */
  avgPricePerMt: number | null;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  LISTING_CREATED: 'LISTING_CREATED',
  ORDER_EXECUTED: 'ORDER_EXECUTED',
  ORDER_SETTLED: 'ORDER_SETTLED',
  EWR_INGESTED: 'EWR_INGESTED',
  EWR_ENCUMBERED: 'EWR_ENCUMBERED',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  commodityType?: string | null;
  /** @nullable */
  valueUsd?: number | null;
  createdAt: string;
}

export interface WarehouseStat {
  warehouseCode: string;
  ewrCount: number;
  totalWeightMt: number;
  activeCommodities?: string[];
}

export type ListEwrsParams = {
ownerId?: number;
state?: ListEwrsState;
commodityType?: ListEwrsCommodityType;
};

export type ListEwrsState = typeof ListEwrsState[keyof typeof ListEwrsState];


export const ListEwrsState = {
  INGESTED: 'INGESTED',
  MARKET_LISTED: 'MARKET_LISTED',
  LOCK_TRADING: 'LOCK_TRADING',
  SETTLED: 'SETTLED',
  ENCUMBERED: 'ENCUMBERED',
} as const;

export type ListEwrsCommodityType = typeof ListEwrsCommodityType[keyof typeof ListEwrsCommodityType];


export const ListEwrsCommodityType = {
  MAIZE: 'MAIZE',
  RICE: 'RICE',
  COFFEE: 'COFFEE',
  TEA: 'TEA',
  AVOCADO: 'AVOCADO',
} as const;

export type ListSpotListingsParams = {
commodityType?: ListSpotListingsCommodityType;
grade?: string;
warehouseCode?: string;
minPrice?: number;
maxPrice?: number;
status?: ListSpotListingsStatus;
};

export type ListSpotListingsCommodityType = typeof ListSpotListingsCommodityType[keyof typeof ListSpotListingsCommodityType];


export const ListSpotListingsCommodityType = {
  MAIZE: 'MAIZE',
  RICE: 'RICE',
  COFFEE: 'COFFEE',
  TEA: 'TEA',
  AVOCADO: 'AVOCADO',
} as const;

export type ListSpotListingsStatus = typeof ListSpotListingsStatus[keyof typeof ListSpotListingsStatus];


export const ListSpotListingsStatus = {
  ACTIVE: 'ACTIVE',
  LOCKED: 'LOCKED',
  SETTLED: 'SETTLED',
  CANCELLED: 'CANCELLED',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  PENDING_SETTLEMENT: 'PENDING_SETTLEMENT',
  SETTLED: 'SETTLED',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};