# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WRS Digital Agricultural Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ewrs:
    get:
      operationId: listEwrs
      tags: [ewrs]
      summary: List electronic warehouse receipts
      parameters:
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
            enum: [INGESTED, MARKET_LISTED, LOCK_TRADING, SETTLED, ENCUMBERED]
        - name: commodityType
          in: query
          schema:
            type: string
            enum: [MAIZE, RICE, COFFEE, TEA, AVOCADO]
      responses:
        "200":
          description: List of eWRs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ewr"
  /ewrs/{ewrId}:
    get:
      operationId: getEwr
      tags: [ewrs]
      summary: Get electronic warehouse receipt by ID
      parameters:
        - name: ewrId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: eWR detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ewr"
        "404":
          description: Not found
...
  /listings:
    get:
      operationId: listSpotListings
      tags: [listings]
      summary: List spot market listings
      parameters:
        - name: commodityType
          in: query
          schema:
            type: string
            enum: [MAIZE, RICE, COFFEE, TEA, AVOCADO]
        - name: grade
          in: query
          schema:
            type: string
        - name: warehouseCode
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, LOCKED, SETTLED, CANCELLED]
      responses:
        "200":
          description: List of spot listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpotListing"
    post:
      operationId: createSpotListing
      tags: [listings]
      summary: Create a new spot listing (Producer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotListing"
        "400":
          description: Bad request
        "403":
          description: Forbidden
  /listings/{listingId}:
    get:
      operationId: getSpotListing
      tags: [listings]
      summary: Get spot listing by ID
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing detail with eWR metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotListingDetail"
        "404":
          description: Not found
    delete:
      operationId: cancelSpotListing
      tags: [listings]
      summary: Cancel a spot listing (seller only)
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotListing"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    Ewr:
      type: object
      required: [id, ewrsReceiptId, wrscSignature, warehouseCode, commodityType, grade, weightMt, moisturePct, harvestSeason, isLienActive, state, ownerId, issuedAt]
      properties:
        id:
          type: integer
        ewrsReceiptId:
          type: string
        wrscSignature:
          type: string
        warehouseCode:
          type: string
        commodityType:
          type: string
          enum: [MAIZE, RICE, COFFEE, TEA, AVOCADO]
        grade:
          type: string
        weightMt:
          type: number
        moisturePct:
          type: ["number", "null"]
        harvestSeason:
          type: string
        isLienActive:
          type: boolean
        lienHolderId:
          type: ["integer", "null"]
        state:
          type: string
          enum: [INGESTED, MARKET_LISTED, LOCK_TRADING, SETTLED, ENCUMBERED]
        ownerId:
          type: integer
        ownerName:
          type: ["string", "null"]
        expiryAt:
          type: ["string", "null"]
          format: date-time
        issuedAt:
          type: string
          format: date-time
        estimatedValueUsd:
          type: ["number", "null"]
...
    SpotListing:
      type: object
      required: [id, ewrId, sellerId, pricePerMt, currency, status, createdAt]
      properties:
        id:
          type: integer
        ewrId:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: ["string", "null"]
        pricePerMt:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [ACTIVE, LOCKED, SETTLED, CANCELLED]
        commodityType:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        weightMt:
          type: ["number", "null"]
        warehouseCode:
          type: ["string", "null"]
        moisturePct:
          type: ["number", "null"]
        harvestSeason:
          type: ["string", "null"]
        totalValueUsd:
          type: ["number", "null"]
        platformFeeUsd:
          type: ["number", "null"]
        escrowFeeUsd:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    SpotListingDetail:
      type: object
      required: [listing, ewr]
      properties:
        listing:
          $ref: "#/components/schemas/SpotListing"
        ewr:
          $ref: "#/components/schemas/Ewr"
    SpotListingInput:
      type: object
      required: [ewrId, pricePerMt, currency]
      properties:
        ewrId:
          type: integer
        pricePerMt:
          type: number
        currency:
          type: string
          default: USD
...