# lib/api-zod/src/generated/types/spotListingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WRS Digital Agricultural Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SpotListingInput {
  ewrId: number;
  pricePerMt: number;
  currency: string;
}